Redefined how base trampolines for function exit/return points are created, inserted, executed, etc.  We are unable to always find (all of the) return branches out of a function.  So intead, we modified how things are done a bit.
By modifying the entry base trampoline of a function, and allocating a
few instructions, we are able to set the link register (which points to
the instruction where the pc will go to after it exits the function) to
point to the return base trampoline.  So, when the function exits it will
jump to the return base trampoline, which will reset the link register to
point to where it should really go--the original value of the link register,
where the function should really go after it exits--is stored on the stack,
in the callee's link-area word #4.

In order to get things working, and be a bit conservative, whenever an entry
or exit trampoline is created, the other one will also be created.  So you
will always have both, and entry and exit trampoline for a function.  This
is done to prevent us from going in and changing the entry trampoline when
an exit trampoline is added.