This breaks apart the behavior we use when searching for a function by address. It's actually a bad solution, but it works _FOR NOW_.
When we determine the address of a function, there are multiple
possible answers. The first is the offset the function is at within a
library. The second, given a running process, is the function's
address in that process' memory space. The third is the (possibly
relocated) address of the function.

Here's the problem. The function pd_Function::getAddress is supposed
to return the third value, but defaults to the first if the function
is not relocated. The two different values should really be returned
by different functions.

We were having problems finding a function by address because of this
behavior. Multiple matches were being returned because a single
address matched one function's absolute address (2 above) and another
function's offset (1 above). The offset should never have been
consulted, realistically.

Fix: check by address type 2 above, and then by address 3. NEVER check
by address 1 (which, unfortunately, we still do).