This commit fixes a bug in function relocation pertaining to the creation of new instPoints after a function has been relocated, on sparc. The problem was that when new instPoints for the rewritten function were being created, the constructor was checking the original function to find out what instructions were located around the instrumentation point. (This knowledge is used to determine how to instrument the point). Instead, the constructor should have been examining the rewritten function.
I added a new constructor to the instPoint class with a different parameter
list that examines the instructions in the rewritten function. I also changed
the function newCallPoint() so that rather than generating a new instPoint
object, an instPoint object is created before newCallPoint is called and that
instPoint is then passed into the function as a parameter.