M dyninstAPI/src/aix.C M dyninstAPI/src/inst-power.C M dyninstAPI/src/process.C M dyninstAPI/src/symtab.h
This commit fixes the AIX stack walk behavior and merges the threaded and
non-threaded stack walk code. For most walks we can pull all required
data from the stack, and so it doesn't matter if we're interested in
a kernel thread, pthread, or process because we're reading from memory.

In addition, this fixes bad inferior RPC behavior for multiple threads.

The old behavior incorrectly read the link register value from a register
when we were in the uppermost (leaf) function, which is only correct
if the function does not make calls.

The new behavior is as follows:

  When a function is parsed, if the first instruction is a MFLR (move
  from link register) then the function is labelled as a non-leaf function.
  Similarly, if it pushes the stack pointer down, then it is labelled
  as creating a stack frame.

  When the stack walk is done, the next LR and FP are grabbed from the
  stack (default behavior). However, if the function is a leaf or does
  not make its own stack frame, we handle it specially. For a leaf function,
  the LR is taken from the register, not its storage location. If the
  function does not create its own stack frame, the same FP value is
  reused.

  Finally, we detect whether the LR points at our exit tramp instead of
  the caller. If this is the case, we correct it.

This change has added a member variable to the function_base class:
makesNoCalls (is a leaf function). This is only used on AIX for now.

aix.C: fixed getCallerFrame{Normal|Thread|LWP}
inst-power.C: Mark function while parsing
process.C: Fix multi-thread RPC handling
symtab.h: add member variable