This commit fixes a bug in process::insertTrapAtEntryPointOfMain(). There seems to be a problem with the first SIGTRAP after a successful exec call. At the time of the first SIGTRAP, any attempt to read or write the child data space fails.
If the child is instructed to continue, it will eventually stop
in a useable state (before the first instruction of main).  However,
a SIGTRAP will *NOT* be generated on the second stop.  PROCFS also
stops in a strange state (prstatus_t.pr_info.si_code == 0).

Looks like this code was in place before.  I don't know why it was
removed.

Ray Chen