This commit fixes a bug I introduced in my previous commit.  Namely, the libdyninstAPI_RT.so would not build correctly with cc because _init was multiply-defined.
The problem is the runtime library now needs
to have a function called upon loading
the library.  There is no compiler independent
way to code this into the source.

I have added a flag to the solaris Makefile
to determine if you are compiling with cc or gcc.
If you are compiling with cc the appropriate
cc pragma is used.  If you are using gcc the
_init function and -nostdlib flag is used as
before.