Folks: 	This commit brings a temporary fix to the problem with shared libraries moving around in mutated (via save the world) binaries.  The mutator currently throws an error to the user that they can optionally ignore when it sees dlopen being used.
	Now with this commit the mutated binary
will check each loaded shared library to make sure
that if it is instrumented it is loaded in the
correct location.  If it is not loaded in the correct
location an error message is printed and the
application exits.  There is no recoverable way
to handle this right now, the instrumentation will
simply be jumping to the wrong addresses.

	In the future I hope to fix this, that is
to say have the mutated binary patch its own
mutations after the shared library is loaded to
take into account the new addresses.  This will
be a large and difficult task because of the way
we generate code to place in the mutatee.  Note
that to correctly patch the mutated binary every
instruction the mutator inserts into the mutatee
that touches memory or performs flow control, either
through relative or absolute addresses, must be
looked at to ensure correctness.  That is a lot
of instructions and they are generated all over
the place.

	Note this commit is for Solaris and Linux.

	IMPORTANT: This commit also moves the save the
world functionality to RTmutatedBinary.c.  This is part
of an effort to pull the Dyninst and Paradyn runtime
shared library apart and create two separate,
non-interdependent objects.

cheers,c
chadd