This update allows the user to load a shared library into the mutated binary when save the world is used.
BPatch_thread::loadLibrary is given another parameter,
a bool defaulted to false.  If this is set to true the
library is reloaded when the mutated library is loaded.
If this is set to false (default) it is not reloaded by
the mutated binary unless the mutated binary unless it
is part of the normal execution sequence to do so.

This control is given to the user because it could be the
case that loading a shared library can change the execution
of an executable by redefining a function.