Folks: 	This commit fixes a number of bugs in the save the world code.
1. The mutated binary is now put in a subdirectory
created under the current working directory. Previously
it was put in a subdirectory created under the directory
holding the mutatee.  This was a problem when saving
instrumentation placed in a mutatee in a non-writeable
directory (ie java).

2. If the mutatee was compiled with a compiler other
than gcc the _DYNAMIC table could have been created
such that it was impossible (with the implentation
of save the world) to add the DyninstAPI runtime
shared library to the dependency list.  The method
for adding the shared library has been changed to a
much more general way.  This could still fail if there
is not enough room between the text and data segments
to stick in the new _DYNAMIC table.  This is highly
unlikely because the gap is ~1 page and the _DYNAMIC
table is usually less than 250 bytes. This limitation
comes from the implementation of the operating
system's loader.


enjoy
chadd

M dyninstAPI/src/addLibrary.C
M dyninstAPI/src/addLibrary.h
M dyninstAPI/src/addLibraryLinux.C
M dyninstAPI/src/addLibraryLinux.h
M dyninstAPI/src/process.C
M dyninstAPI/src/writeBackElf.C