Numerous shamm bugfixes.
#371  emitFuncCall broken for certain C++ mutatees

the call to findOnlyOneFunction in emitFuncCall() would fail
when the callee function was an overloaded C++ function (more than
one pretty name match would be found).

Solution:  change AST function call nodes to use symTabName() exclusively.
Also changed the function signature of emitFuncCall so that it no longer
takes either callee name or function_base as an argument.  Instead it
takes the callee's base address directly.  All symtab lookups must now
therefore be done before calling emitFuncCall().  This was done because
the symtab lookup code should be platform independent, and we had 4 or 5
slightly different versions of the symbol lookup going on in each of the
different platform-specific versions of emitFuncCall().  Should be
cleaner this way.

#370   BPatch_point::getCalledFunction fails on C++ mutatee

getCalledFunction() uses findAllFuncsByName(), which used to only look
at pretty names.  Now, if a pretty name lookup fails, it checks in the
mangled pile as a last ditch beore failing.

#365   findFunction segfaults on mangled-name argument

Fixed logic error in the way we were checking for duplicate mangled names
when building up the funcsByMangled hash.  In fact, I just got rid of
mangled function duplicate checking altogether.  It should not be
necessary.

#361   findFunction does not find dlopen

Thanks to Vic for finding that online gnu symbol versioning reference.
This implements a 0th order "demangling" scheme whereby default symbols
 (of the form <symbol>@@<version>) are given the "demangled" name <symbol>
as well.