This commit fixes test9 on Solaris 2.8. It uses the /proc file system rather than dladdr() to determine if a memory location is owned by th process.
It also improves the correctness of the mutated binary produced by
Dyninst on Solaris and Linux (but this does not appear to have any
effect on the results of the tests).  the first entry of the
.got table contains the address of the .dynamic section.

the old .dynamic section is now completely zeroed out so that
if anything tries to access it we should get an error message
of some kind quickly.