This fixes test4 subtest 3 on linux
If you relocate a function in a shared library,
then call exec (WITHOUT CALLING FORK) the function will continue to be
marked as relocated BY THE EXEC'ED PROCESS even though the shared
library will have been reloaded.

So, to fix this, if we are an execed process, we look to see if each
function is marked as relocated by the exec'ed process and remove the
relocation tag connecting it with the said process.

The function unrelocatedByProcess was added to pd_Function in symtab.h

This only effects exec and not fork since fork creates a new process
and exec does not.  check to see how pd_Function::hasBeenRelocated()
works for more info.