(1)  Compilability with gcc 3.4.3 ( wanted to check if compiling gcc with --enable-threads=posix, vs --enable-threads=single made a difference -- our default gcc 3.3.3 compiler is "single", but our gcc3.4.3 is posix ). No discernable difference, except test1 subtest24 fails when dyninst is compiled with gcc 3.4.3.
(2)  Examination of how we generate arbitrary inst points -- the
historical way to do it is to get a basic block and then call
BPatch_image::createInstPointAtAddr() on every address in the block,
expecting failures along the way (particularly on x86).  There was
a comment in test1 suggesting that this was the "second stupidest way to
do it" and that we really need to iterate over instructions.

So...  I added an api:  BPatch_instruction::getInstPoint().  Now, to
generate arbitrary inst points, we can get all the instructions in a basic
block and get the instPoints explicitly, as opposed to doing all of the
previously expected guesswork.

My reasons for doing this was that test1, subtest20 was _crawling_ on
alpha -- for some reason it was generating 1500+ inst points, some of
which (guessing half) were duplicates.

Unfortunately subtest20 is still mind-numbingly slow, simply due to the
time required to generate, install, and link 750+ snippets.  I'm not
familiar enough (yet) with the multiTramp code to know if this can be sped
up, but it seems intractably slow at the moment.