Problem 1:
Historically Dyninst does a sleep(1) after receiving notification of a
fork() exit to allow the OS time to fully create the process.  Somehow
this went missing in the recent re-org of the fork handling code and all
hell broke loose.

Its not the best solution, but I just put it back.

Problem 2:

The mailbox system is quite flexible, what with being able to handle
recursive chains of callbacks, but the downside of this flexibility is
that it requires outside hints as to when it is ok to delete a callback
that has already been executed.

It is bad, for example, to delete a callback that is waiting for the
completion of another callback, or, more likely, to delete a callback when
we are later going to need to query it for its return value.

The memory clobber that we were seeing was actually correct behavior --
from the perspective of the mailbox.  Certain callbacks were being
deleted after being executed.  The error was that these callbacks
should've known better and indicated that it was not safe to delete them
until the desired event had been received.