Fix a bug with multithreaded process control.
We can get into a situation like so:

Event: library load on thread 1
  - assigned to Handler 1
Event: iRPC finish on thread 2
  - assigned to Handler 2
Handler 1 finishes and calls async continue
  - process continues, although handler 2 is still in flight

The new algorithm checks for whether other handler threads are active.
This is normally handled by the signal generator thread, but in many
cases it's waiting in waitpid (and thus not useful).

The real answer is redesign the system to use a single handler thread;
this distributed consensus stuff isn't worth the complexity.