This is still true, but the static data structure is now: map [ Annotatee <--> map [ annotation_type <--> vector<Annotation> ] ]
(The present, simpler implementation is broken and does not allow for the
proper discernment of annotation types, by object).  The discussion of
whether the amount of access time is sufficiently small can commence...

This also adds a first stab at a public interface to annotations -- So far
I have tried to keep in a notion of arbitrary object annotations, ie, you
can say something like "annotate this object A with this B".

Anyways...  to both test this and to provide an example, I applied this
public interface to SymtabAPI::Symbol by replacing the notion of the upPtr
with an externally provided annotation.  In other words, since a Symbol
never needs to have knowledge of an image_func, or an image_variable, this
can be handled completely externally to symtabAPI.  Furthermore this
allows for the distinction of what the effective upPtr points to (an
image_func or an image_variable) -- creating a greater sense of type
safety that is independent of code-context.