This adds some basic infrastructure to allow for varying the underlying implementation of annotations by providing a template parameter.  The template parameter is referred to as IMPL in the Annotatable.h file and is presently an enum:
typedef enum {
   anno_impl_static_sparse,
   anno_impl_static_dense,
   anno_impl_local_dense,
} annotation_implementation_t;

The implementations that I put together are schematic at best, except
for the static_sparse one, which should be functionally the same as
what we had before. The
last one may not strictly be an "annotation" scheme since all storage
used by it is local to the annotatee, not static, and not really tied into
the annotation naming scheme.